requirements.txt
aniso8601==8.1.1
argon2-cffi==20.1.0
asgiref==3.3.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupPy==1.14
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.4
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.9
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
prometheus-client==0.9.0
prompt-toolkit==3.0.14
psutil==5.8.0
psycopg2-binary==2.8.4
pycparser==2.20
pydevd===pycharm
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==22.0.2
qtconsole==5.0.2
QtPy==1.9.0
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
tablib==2.0.0
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xlrd==1.2.0
xlwt==1.3.0
